Feature/exam explain result and genereate new practice (#73)
* Maintain: Refactor code

* Exam: Explain wrong answer and grade essay question

+++ src/main/java/com/cabybara/prolearningplatform/dto/response/exam/EssayGradingResponseDto.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record EssayGradingResponseDto(
         @JsonProperty("attemptId") Long attemptId,
         @JsonProperty("questionId") Long questionId,
-        @JsonProperty("score") Integer score,
+        @JsonProperty("score") Double score,
         @JsonProperty("maxPoints") Integer maxPoints,
         @JsonProperty("feedback") String feedback
 ) {}
